+++ game/city/city.cpp
@@ -96,7 +96,7 @@ City::City(Framework &fw, std::string mapName)
 	std::list<Tile*> path;
 	path = this->findShortestPath(Vec3<int>{0,0,9}, Vec3<int>{99,99,9});
 
-	LogInfo("Route found in %zu steps", path.size());
+	LogInfo("Route found in %u steps", (unsigned int)path.size());
 	for (auto tile : path)
 	{
 		LogInfo("Tile {%d,%d,%d,}", tile->position.x, tile->position.y, tile->position.z);
